feat: 将设置的样式抽离出来
feat: 将设置的样式抽离出来
